+++ CHANGELOG.md
@@ -5,6 +5,12 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- `findClosestTarget` function
+### Changed
+- `processAudio` function modified
+- `updateDisplay` function modified
+- Auto mode bug fixed
 ## [0.2] - 2026-09-03
 ### Added
 - `setStatus` function